requirements.txt
streamlit>=1.28.0
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
lightgbm>=4.1.0
xgboost>=2.0.0
plotly>=5.18.0
tensorflow-probability>=0.21.0
pyro-ppl>=1.8.6
openpyxl>=3.1.2  # for Excel file support
python-dotenv>=1.0.0
pytest>=7.4.0  # for testing
optuna>=3.5.0  # for hyperparameter tuning 